fixed user deleting self bug

+++ backend/routes/users.js
@@ -46,16 +46,24 @@ router.delete('/delete/:id/:page', authUser, (req, res, next) => {
     const { page } = req.params;
     const limit = 10;
     const adminAccount = req.user.email;
-    deleteUser(id).then((deletedUser) => {
-        renderDeletedAccountEmail(adminAccount).then((template) => {
-            // Would enter "deletedUser.email" if the email API would be in live use
-            sendDeletedEmail(template, 'ricotalvar@pohi.io').then(() => {
-                findAllUsers(limit, page, limit * page).then((userData) => {
-                    res.json({ users: userData.inRange, hasMore: userData.hasMore, page });
+    // If user is trying to delete themselves just send back all the accounts and do nothing
+    if (req.user._id !== id) {
+        deleteUser(id).then((deletedUser) => {
+            renderDeletedAccountEmail(adminAccount).then((template) => {
+                // Would enter "deletedUser.email" if the email API would be in live use
+                sendDeletedEmail(template, 'ricotalvar@pohi.io').then(() => {
+                    findAllUsers(limit, page, limit * page).then((userData) => {
+                        res.json({ users: userData.inRange, hasMore: userData.hasMore, page });
+                    }).catch(next);
                 }).catch(next);
             }).catch(next);
         }).catch(next);
-    }).catch(next);
+    }
+    else {
+        findAllUsers(limit, page, limit * page).then((userData) => {
+            res.json({ users: userData.inRange, hasMore: userData.hasMore, page });
+        }).catch(next);
+    }
 });
